02-image_detection/00-cv_mser/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.2.0.32
PyExecJS==1.5.1
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
